Fix installation of Google Cloud SDK package repo
The key currently available for download is ASCII-armoured, so needs the
.asc extension rather than .gpg – otherwise you get warnings like:

    The key(s) in the keyring /etc/apt/trusted.gpg.d/gcloud.gpg
    are ignored as the file has an unsupported filetype.

and ultimately the task fails, with:

    The following signatures couldn't be verified because the public key
    is not available: NO_PUBKEY B53DC80D13EDEF05

It's not clear what the correct solution here (which would cope with
either ASCII-armoured or binary keys) would be – particularly since
apt-key is deprecated, so piping to apt-key and getting it to handle the
problem isn't really an option. For now, let's just get it working.